feat: adjust store dir
