+++ src/store/cart/cartStore.ts
@@ -1,32 +1,4 @@
-import type { CartType, Product } from '@/types';
-
-export const tokenCheck = async (
-  token: string | null
-): Promise<Response | undefined> => {
-  const data = await fetch('/api/auth/email/verify', {
-    method: 'PUT',
-    headers: {
-      'Content-type': 'application/json',
-    },
-    body: JSON.stringify({ token }),
-  })
-    .then((res) => {
-      if (!res.ok) {
-        return res;
-      }
-      return res.json();
-    })
-    .catch((error) => error);
-
-  return data;
-};
-
-export const getProducts = async (): Promise<Product[]> => {
-  const dataList: Product[] = await fetch('/api/get-products').then((res) =>
-    res.json()
-  );
-  return dataList;
-};
+import type { CartType } from '@/types';
 
 export const getCart = async (): Promise<CartType> => {
   const data = await fetch('/api/cart')
