+++ src/store/products/productStore.ts
@@ -1,0 +1,8 @@
+import { Product } from '@/types';
+
+export const getProducts = async (): Promise<Product[]> => {
+  const dataList: Product[] = await fetch('/api/get-products').then((res) =>
+    res.json()
+  );
+  return dataList;
+};
